refactor dtos
