+++ apps/frontend/app/components/AddComment.tsx
@@ -1,9 +1,9 @@
 'use client';
 import React, { useState } from 'react';
 import { Rating, Stack } from '@mui/material';
-import { SendCommentDto } from '../Dtos';
 import { useSession } from 'next-auth/react';
 import { RequestNextNest } from '@frontendlibs';
+import { SendCommentDto } from './dtos';
 
 type Props = {
     restaurant_id: string;
