+++ apps/frontend/app/components/Comments.tsx
@@ -1,6 +1,5 @@
 'use client';
 import React, { useState } from 'react';
-import { CommentDetails } from '../Dtos/CityDetails.dto';
 import { Stack } from '@mui/material';
 import Rating from '@mui/material/Rating';
 import Modal from './Modal/Modal';
@@ -9,6 +8,7 @@ import DropdownThreedots from './Dropdown/Dropdown';
 import { useSession } from 'next-auth/react';
 import { UpdateCommentDataDto } from './dtos/navigate.type';
 import { RequestNextNest } from '@frontendlibs';
+import { CommentDetails } from './dtos';
 
 export const Comments = ({
     comments: comment
