+++ apps/frontend/app/components/HomePage.tsx
@@ -8,7 +8,7 @@ import {
 } from '@react-google-maps/api';
 import { useState } from 'react';
 import '../App.css';
-import { CitiesJsonDto } from '../Dtos';
+import { CitiesJsonDto } from './dtos';
 import Link from 'next/link';
 
 const App = ({ cities }: { cities: Array<CitiesJsonDto> }) => {
