+++ apps/frontend/app/components/dtos/CityDetails.dto.ts
