+++ apps/frontend/app/components/dtos/index.ts
@@ -1,3 +1,5 @@
 export * from './Cities.dto';
 export * from './CityDetails.dto';
+export * from './Login.dto';
+export * from './navigate.type';
 export * from './SendCommentDto';
